+++ BlueprintUICommonControls/Sources/SegmentedControl.swift
@@ -12,8 +12,19 @@ public struct SegmentedControl: Element, Measurable {
     public var font: UIFont = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.body)
     public var roundingScale: CGFloat = UIScreen.main.scale
 
-    public init(items: [Item] = []) {
+    public init(items: [Item] = [], configure: (inout SegmentedControl) -> Void = { _ in }) {
         self.items = items
+        configure(&self)
+    }
+
+    public init(
+        selection: Selection = .none,
+        font: UIFont = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.body),
+        @Builder<Item> itemBuilder: () -> [Item]
+    ) {
+        items = itemBuilder()
+        self.selection = selection
+        self.font = font
     }
 
     public mutating func appendItem(title: String, width: Item.Width = .automatic, onSelect: @escaping () -> Void) {